CC-1177: Fixed CompanyPage controllers

+++ Bundles/CompanyPage/src/SprykerShop/Yves/CompanyPage/Controller/CompanyRoleController.php
@@ -12,11 +12,12 @@ use Generated\Shared\Transfer\CompanyRoleCriteriaFilterTransfer;
 use Generated\Shared\Transfer\CompanyRoleResponseTransfer;
 use Generated\Shared\Transfer\CompanyRoleTransfer;
 use Generated\Shared\Transfer\CompanyUserCollectionTransfer;
-use Generated\Shared\Transfer\CompanyUserTransfer;
 use Generated\Shared\Transfer\PermissionCollectionTransfer;
+use SprykerShop\Yves\CompanyPage\Form\CompanyRoleForm;
 use SprykerShop\Yves\CompanyPage\Plugin\Provider\CompanyPageControllerProvider;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class CompanyRoleController extends AbstractCompanyController
 {
@@ -71,6 +72,8 @@ class CompanyRoleController extends AbstractCompanyController
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
      *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+     *
      * @return array
      */
     protected function executeDetailsAction(Request $request): array
@@ -81,6 +84,10 @@ class CompanyRoleController extends AbstractCompanyController
         $companyRoleTransfer = $this->getFactory()
             ->getCompanyRoleClient()
             ->getCompanyRoleById($companyRoleTransfer);
+
+        if (!$this->isCurrentCustomerRelatedToCompany($companyRoleTransfer->getFkCompany())) {
+            throw new NotFoundHttpException();
+        }
 
         $companyRolePermissions = $this->getFactory()
             ->getCompanyRoleClient()
@@ -99,6 +106,8 @@ class CompanyRoleController extends AbstractCompanyController
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
      *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+     *
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
      */
     public function deleteAction(Request $request): RedirectResponse
@@ -106,6 +115,14 @@ class CompanyRoleController extends AbstractCompanyController
         $idCompanyRole = $request->query->getInt('id');
         $companyRoleTransfer = new CompanyRoleTransfer();
         $companyRoleTransfer->setIdCompanyRole($idCompanyRole);
+
+        $companyRoleTransfer = $this->getFactory()
+            ->getCompanyRoleClient()
+            ->getCompanyRoleById($companyRoleTransfer);
+
+        if (!$this->isCurrentCustomerRelatedToCompany($companyRoleTransfer->getFkCompany())) {
+            throw new NotFoundHttpException();
+        }
 
         $companyRoleResponseTransfer = $this->getFactory()->getCompanyRoleClient()->deleteCompanyRole($companyRoleTransfer);
 
@@ -121,6 +138,8 @@ class CompanyRoleController extends AbstractCompanyController
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
      *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+     *
      * @return \Spryker\Yves\Kernel\View\View|\Symfony\Component\HttpFoundation\RedirectResponse
      */
     public function confirmDeleteAction(Request $request)
@@ -134,6 +153,10 @@ class CompanyRoleController extends AbstractCompanyController
             ->getCompanyRoleClient()
             ->getCompanyRoleById($companyRoleTransfer);
 
+        if (!$this->isCurrentCustomerRelatedToCompany($companyRoleTransfer->getFkCompany())) {
+            throw new NotFoundHttpException();
+        }
+
         if ($companyRoleTransfer->getIsDefault()) {
             $this->addErrorMessage(static::ERROR_MESSAGE_DEFAULT_COMPANY_ROLE_DELETE);
 
@@ -145,6 +168,8 @@ class CompanyRoleController extends AbstractCompanyController
 
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
+     *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
      *
      * @return \Spryker\Yves\Kernel\View\View
      */
@@ -158,6 +183,10 @@ class CompanyRoleController extends AbstractCompanyController
         $companyRoleTransfer = $this->getFactory()
             ->getCompanyRoleClient()
             ->getCompanyRoleById($companyRoleTransfer);
+
+        if (!$this->isCurrentCustomerRelatedToCompany($companyRoleTransfer->getFkCompany())) {
+            throw new NotFoundHttpException();
+        }
 
         $viewData = [
             'idCompanyRole' => $idCompanyRole,
@@ -201,7 +230,7 @@ class CompanyRoleController extends AbstractCompanyController
             ->handleRequest($request);
 
         if ($companyRoleForm->isSubmitted() === false) {
-            $idCompany = $this->getCompanyUser()->getFkCompany();
+            $idCompany = $this->findCurrentCompanyUserTransfer()->getFkCompany();
             $companyRoleForm->setData($dataProvider->getData($idCompany));
         }
 
@@ -238,6 +267,8 @@ class CompanyRoleController extends AbstractCompanyController
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
      *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+     *
      * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
      */
     protected function executeUpdateAction(Request $request)
@@ -255,8 +286,13 @@ class CompanyRoleController extends AbstractCompanyController
         $idCompanyRole = $request->query->getInt(static::PARAMETER_ID_COMPANY_ROLE);
 
         if ($companyRoleForm->isSubmitted() === false) {
-            $idCompany = $this->getCompanyUser()->getFkCompany();
-            $companyRoleForm->setData($dataProvider->getData($idCompany, $idCompanyRole));
+            $idCompany = $this->findCurrentCompanyUserTransfer()->getFkCompany();
+            $data = $dataProvider->getData($idCompany, $idCompanyRole);
+
+            if (!$this->isCurrentCustomerRelatedToCompany($data[CompanyRoleForm::FIELD_FK_COMPANY])) {
+                throw new NotFoundHttpException();
+            }
+            $companyRoleForm->setData($data);
         }
 
         if ($companyRoleForm->isSubmitted() && $companyRoleForm->isValid()) {
@@ -281,7 +317,7 @@ class CompanyRoleController extends AbstractCompanyController
     {
         $criteriaFilterTransfer = new CompanyRoleCriteriaFilterTransfer();
 
-        $criteriaFilterTransfer->setIdCompany($this->getCompanyUser()->getFkCompany());
+        $criteriaFilterTransfer->setIdCompany($this->findCurrentCompanyUserTransfer()->getFkCompany());
         $criteriaFilterTransfer->setPagination($this->createPaginationTransfer($request));
         $criteriaFilterTransfer->setFilter($this->createFilterTransfer(static::COMPANY_ROLE_SORT_FIELD));
 
@@ -354,21 +390,5 @@ class CompanyRoleController extends AbstractCompanyController
         return $this->getFactory()
             ->getCompanyRoleClient()
             ->findNonInfrastructuralCompanyRolePermissionsByIdCompanyRole($companyRoleTransfer);
-    }
-
-    /**
-     * @return \Generated\Shared\Transfer\CompanyUserTransfer|null
-     */
-    protected function findCurrentCompanyUserTransfer(): ?CompanyUserTransfer
-    {
-        $currentCustomerTransfer = $this->getFactory()
-            ->getCustomerClient()
-            ->getCustomer();
-
-        if (!$currentCustomerTransfer) {
-            return null;
-        }
-
-        return $currentCustomerTransfer->getCompanyUserTransfer();
     }
 }
